demo/requirements.txt
gunicorn
whitenoise
pytest
pytest-cov
pytest-django
pytest-factoryboy
